Move parse_human_friendly_id into parsers.py
